stm32f7xx: HAL and LL SD driver fixes
Differentiate between the MMC CMD3 (SET_RELATIVE_ADDR) and the SDIO CMD3
(SEND_RELATIVE_ADDR).

The hal_mmc driver had two bugs:
1. It uses the SEND_RELATIVE_ADDR command, which relies on the card
assigning itself an RCA and returning that in R6. However, the JEDEC MMC
standard has the host assign the RCA and send it as part of CMD3. The
result of CMD3 is in R1, not R6.
2. After the RCA has been assigned, the card must be selected before any
command that requires a data transfer (e.g. CMD8 to get the extended
CSD). If the card is not selected, any command that involves a data
transfer will timeout and no data will be transferred.

This proposal provides two alternative versions of the low level CMD3
function. `SDMMC_CmdSetRelAdd` (old name but new parameters) supports
the MMC protocol, while `SDMMC_CmdSendRelAdd` is provided to support the
SD protocol. However, this changes the published Low Level
specification.