SCF-583: stm32cube: Support CONFIG_USE_UPSTREAM_CUBE
Instead of using the version of the Cube code shipped with Zephyr, allow for
using the upstream one.

This depends on having an environment variable STM32CUBEF7_BASE pointing to the
upstream checkout; it updates the src and include paths and copies
stm32f7xx_hal_conf.h to a separate directory so we can use it without getting
the rest.

If CONFIG_USE_UPSTREAM_CUBE is not enabled, nothing changes.