SCF-885 SCF-859 hal_mmc: block interrupts while setting up DMA
While not immediately obvious why this is needed, when both the SAI and
MMC drivers are being setup on DMA2 it is possible to corrupt the IP
(intra-process) register and get a crash resulting. This crash _always_
occurs at the `MODIFY_REG` call for the stream control register, with an
unaligned access violation. Block interrupts around this call as a stop
gap fix.