SCF-1137 merge Logi update and latest STM HAL
This reworks an earlier commit that broke whitespace unnecessarily, and
includes a minor improvement to the cache initialization to ensure that
the cached extCSD is up to date. While this introduces no new
functionality, it does mean that a log message during boot that was
previously incorrect (claiming the cache was disabled) now prints the
truth.